Enable weight sharing between threads.
This commit changes the way we save information in the context. It removes the
TrainableData class and instead makes the network operations handle tensors
manually. It adds the concept of shared tensors that are owned by the first
thread (ctx0). As a side effect this reduces the noise in the initialization of
the weights because the random method only initializes the first context, so we
get the same initialization weights no matter how many threads we have.  This
work is derived by the task of loading the caffe model.