Add the AvgPool operator.
Implement the AvgPool operation (as defined in https://arxiv.org/abs/1312.4400). AvgPool is implemented as a mode of operation for the existing MaxPool node.
In the future we may want to split this operation out into a dedicated node because they share almost no code.

This test also adds a grad check unit test.