Fix a bug in the implementation of divergent networks.
With this change we move the responsibility for clearing the gradient that we need to update in the backdrop phase from the backdrop into the network. We prepare for the next backprop iteration by zeroing the gradient tensors during the forward() loop. Notice that we only zeros the temporary grad tensors that match the output tensors but not the gradient tensors that are paired with filters. These are cleared during the learning process at the end of the batch.