Change we way we update tensors while training.
This is the first step of implementing batched tensors (batched inference and training). This change renames ArrayNode into Variable, and removes the concept of updateInputs for general nodes. SoftMax and RegressionNode are now accepting the Variable vectors as inputs. The trainer is now only able to train Variables. Variables are typed, and can be any type, unlike ArrayNode, that were floats only.