Add a configuration that limits the number of threads.
Summary:
Add a configuration that limits the number of threads. The number of
threads affects the frequency of weight updates because each thread holds some
gradient tensor, and then the tensors update the state at once. This change also
reduces the number of thraeds in the tests to 1, to make sure that the tests are
predictable on all machines.

Test Plan: Run the tests on a machine with lots of cores. Rely on existing tests.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: jongsoo, #glow

Differential Revision: https://phabricator.intern.facebook.com/D5690697

Signature: t1:5690697:1503515581:1aa966c30229d8ee36814838e8a4a431ec74461e

Blame Revision: None