Fix the Glow build by fixing the caffe2 protobuf path issue.
Summary:
Fix the Glow build by fixing the caffe2 protobuf path issue. The
problem was that when the header files that were generated for the caffe proto
description file were not generated in the right path. The problem was not
discovered because an old header file remained in the build directory.

The solution to the problem was to move all of the dependencies on the Caffe2
header file into the src directory and keep the header clean of any C2
datastructures.

Test Plan: Tested the build in a clean environment.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: azzolini, tulloch, jiayq, #glow

Differential Revision: https://phabricator.intern.facebook.com/D5696462

Signature: t1:5696462:1503549122:4a1e15a749378b13d1cb9f2f080f340bcd7d32a7

Blame Revision: None