Fix a bug in the SoftMax node visitor.
Summary:
Fix a bug in the node visitor. We changed the design of the SoftMaxNode
and made the expected result a parameter of the node. However, we forgot to
change the visitor to visit this operand. This commit makes the visitor visit
the 'expected' operand. The bug was discovered by looking at the dotty network
dump. The image did not display the node.

Test Plan: Tested by inspecting the dotty dump.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: #glow

Differential Revision: https://phabricator.intern.facebook.com/D5705018

Signature: t1:5705018:1503629663:e27133267fb7042593209321e3c68a541952c86e

Blame Revision: None