build: treat googletest as an external project
Summary:
Build googletest as an external project.  This cleans up the targets that we
generate by not treating googletest as part of our codebase even though it
resides in the tree.  It allows us to build googletest separately and use it as
if it were previously installed on the system.

More importantly, this sets the ground work to allow us to use a newer version
of CMake (3.7+) to run clang-tidy in an automated fashion without relying on
custom tooling.  The flags are implicitly handed over to clang-tidy and can be
integrated into the normal build process.

Test Plan:
`cmake -G Ninja -DENABLE_TESTING=YES ..`
`ninja`

Reviewers: nrotem

Reviewed By: nrotem

Differential Revision: https://phabricator.intern.facebook.com/D5734299

Signature: t1:5734299:1504064534:dca446ad41b8c5264b697ac4b15a0ae99974c662