build: remove some variables from Config.h
Summary:
The protobuf dependency is required, the configuration value was never checked.
The `WITH_PNG` is used at one site and inside the build of glow itself.  Pass
that from the command line instead.  Unfortunately, removing the last use of
Config.h is slightly more complicated; leave that for a follow up.

Test Plan: ninja test

Reviewers: #glow, nrotem

Reviewed By: nrotem

Differential Revision: https://phabricator.intern.facebook.com/D5735953