Fix a bug in updateTensor. We were copying from the src tensor into itself.
Summary:
Fix a bug in updateTensor. We were copying from the src tensor into
itself.  Discovered this bug as part of the loader work. Added an assertion and
will add a testcase as soon as we are able to load a whole mode.

Test Plan:
Discovered this bug as part of the loader work. Added an assertion and will add
a testcase as soon as we are able to load a whole mode.

Reviewers: #glow, abdulras, jongsoo

Reviewed By: jongsoo

Subscribers: #glow

Differential Revision: https://phabricator.intern.facebook.com/D5761755

Tags: none

Signature: 5761755:1504335557:bccb9db2c8ddf914a4eb9d24163f1bef408c9ef0

Blame Revision: none