Caffe2 loader: When creating a new variable make sure to register it into the named tensor map.
Summary:
Caffe2 loader: When creating a new variable make sure to register it
into the named tensor map. We will use this functionality when we update the
content of the variable in the loader.cpp test program. If we don't register the
variable then we won't be able to update the contents later.

Test Plan:
We will use this functionality when we update the
content of the variable in the loader.cpp test program. If we don't register the
variable then we won't be able to update the contents later.

Reviewers: #glow, abdulras, jongsoo

Reviewed By: jongsoo

Subscribers: azzolini, jiayq, #glow

Differential Revision: https://phabricator.intern.facebook.com/D5761761

Tags: none

Signature: 5761761:1504335742:c18690617b805db977298a7d96e592129a182195

Blame Revision: none