Teach the loader driver to load png images.
Summary:
Teach the loader program driver to load images in the NCHW format (BGR order)
and pass them for inference in the models that are loaded from disk. This is
just a test program and not a general inference driver.

Test Plan: This is a test. We use it to run VGG19.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: jongsoo, tulloch, #glow

Differential Revision: https://phabricator.intern.facebook.com/D5788800

Tags: none

Signature: 5788800:1504821071:557f7a060163edce34e12961fb13e5035877a227