Load the Resnet50 model.
Summary:
Load the Resnet50 model. This commit loads the resnet50 model. It
implements the deserialization of the batch-normalization node, the average
pooling node, sum-arithmetic node, and fixes the deserialization of the
convolution node that has no bias. This commit also fixes the driver and allows
it to load inputs with the gpu_0 suffix, and loads the last scoring layer and
saves it in the root variable automatically.

Test Plan:
Tested with the resnet50 network. Use existing unit tests for the
different layers.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: jongsoo, azzolini, tulloch, jiayq, #glow

Differential Revision: https://phabricator.intern.facebook.com/D5805600

Tags: none

Signature: 5805600:1505190528:8bfb5f1a861bb947e60a93eb297043c785e9fcc2