Fix the linux build (linkage)
Summary:
Fix the linux build (linkage). The method was defined in the CPP file,
that went into the network Lib. on linux the linkage order matters. By moving
the method to the header file we make sure that the method, that is used outside
of the Network library is not a linkage dependency.

Test Plan: NFC.

Reviewers: #glow, zem

Reviewed By: zem

Subscribers: #glow

Differential Revision: https://phabricator.intern.facebook.com/D5842997

Tags: none

Signature: 5842997:1505499405:eda1dc3203554e395dc5cbec3a3b7c635fd2cf0b