IR: Remove vtable from instructions.
Summary:
This commit adopts LLVM-style isa/cast/dyn_cast. It makes the instructions
non-virtual. This is the first step in implementing the interpreter. Being able
to switch over instruction kind and checking if an instruction is of a specific
kind is the base for the interpreter.

Test Plan: Added a unit test.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: #glow

Differential Revision: https://phabricator.intern.facebook.com/D5849514

Tags: none

Signature: 5849514:1505704055:6fe53439416a66fd8973652b79a8076cadc68b54