IR: implement the rest of the nodes in the IR interpreter.
Summary:
IR: implement the rest of the nodes in the IR interpreter. Added a unit
test to check sigmoid. There are a few untested nodes that will be testes as
part of the next few commits.

Test Plan:
Added a unit test to check sigmoid. There are a few untested nodes
that will be testes as part of the next few commits.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: zem, #glow

Differential Revision: https://phabricator.intern.facebook.com/D5860492

Tags: none

Signature: 5860492:1505923048:91b6f90297e67039914eff22248c49117bb4b65c