build: add LLVM as an external dependency
Summary:
This does a lightweight git clone of LLVM based on the current stable release
into external/llvm.  This is setup work to allow us to pull in more pieces of
LLVM as necessary and use it for eventual code generation.

We exclude the build from the ALL target as we do not wish to build LLVM all the
time.  When the libraries are used, we will wire up the dependencies to ensure
that the build occurs.  This avoids a small penalty on each build (re-running
the cached configure which takes a 1-2s but does not require network connectivty).

Test Plan: ninja

Reviewers: nrotem

Reviewed By: nrotem

Subscribers: #glow

Differential Revision: https://phabricator.intern.facebook.com/D5885799

Signature: 5885799:1506052581:1b21b069ef810159f6bd414340712eb06f6f735e