IR: Port the grad checks to the IR.
Summary:
IR: Port the grad checks to the IR.  We are moving away from the simple
graph interpreter into the IR interpreter. This commit copies the grad checks
from the graph-interpreter into the IR interpreter. This exposed two bugs in the
implementation of the IR interpreter. The first one was that we cleared the
activations and the second one was that we swapped the infer/train methods for
batch normalization. So, having the grad checks is a good idea :)

Test Plan: Added new tests.

Reviewers: #glow, jongsoo

Reviewed By: jongsoo

Subscribers: #glow

Differential Revision: https://phabricator.intern.facebook.com/D5896014

Tags: none

Signature: 5896014:1506142750:6cd89e9524844ec75976c1d41e88e60a9e470e66