Fix incorrect assert in LRNNode::forward
Summary:
This commit fixes an incorrect assert that was added during the
implementation of LRNNode. The intention was to make sure the input
tensor has at least 3 dimensions, but the original assert checked
that the tenor had a minimum depth of 2 instead of a minimal depth
of 1.

Test Plan: All tests pass

Reviewers: jongsoo, nrotem

Reviewed By: nrotem

Differential Revision: https://phabricator.intern.facebook.com/D5907495

Tasks: T21664548

Tags: bootcamp

Signature: 5907495:1506467279:5ee0fee602108554ccb18688b3789be0550172da