Remove redundant includes from headers.
This commit removes the redundant includes from header files. It converts a
bunch of includes into forward declarations and moves files into the right base
library. It also moves the execution engine into a "pimpl" implementation where
the details of the classes are hidden behind a pointer. This can't be done with
a smart pointer that requires the complete type for initialization.