GraphIR: return values into a register.
GraphIR: return values into a register. As part of the migration of the tensors
into the Graph IR we change the way we interact with the execution engine.
Instead of accessing the internals we simply create the 'return' abstraction,
where we have an instruction save the value into a variable that's accessible
from the user code. The user code can no longet access the internal weights of
the interpreter.