Rename NodeVisitor into NodeWalker
The current NodeVisitor visits not only the node, but also recursively visits its children. Thus it is actually a graph walker.

This renaming frees the name NodeVisitor for another type, which would really only visit the nodes and perform some user-defined actions. This type will be added in the subsequent commits.