loader.cpp: improve command line argument parsing
This change adds the llvm CommandLine library and changes the API for
loader.

The loader program now accepts a list of images to process, as well as
specifying the image_mode and caffe2 model paths with command line
flags.

Here are a few example command line invocations:

./loader -image_mode=0to1 -d=resnet50 tests/images/zebra_340.png
./loader -image_mode=0to1 -n=resnet50/predict_net.pb -w=resnet50/init_net.pb tests/images/zebra_340.png

More information can be found by running
./loader -help