build: remove external dependencies
Use LLVM from the system and gtest from LLVM.  This allows us to remove
the external builds and will allow us to integrate glow as a project in
the LLVM tree.

CMake: pull in gtest as a submodule, and rely on an external build of LLVM.

Add a script to build LLVM from source locally.

Remove the code that builds LLVM as ExternalProject_Add.

Fix a bug in the LLVM build script.

Cleanup the pull request.