loader.cpp: refactor the command line options and add loader.h
I was reading some example uses of the command line library on github and
it looks like changing to an opts namespace is more conventional.  I
also needed to add a header file as future commits will use command line
values throughout the file and I do not want to worry about code
placement above/below the cpp definition.  This should also allow for
codesplitting of loader.cpp in the future if necessary.