GraphOptimizer.cpp: add sink optimizations for sigmoid and tanh
Sigmoid and Tanh are types of activation functions commonly used in
Machine Learning.  In this commit, we want to to optimize our graph that
we generate to sink the Sigmoid and Tanh nodes below the RELU nodes.

This is similar to an already existing optimization that we have with
Transpose nodes.

Also included are some tests for these two optimizations.

Ori Shalev originally wrote many of the tests for our optimizations and
I want to point out a very useful debugging technique he mentioned
for debugging our optimizations is to add G.dumpDAG("before.dot") and
G.dumpDAG("after.dot") around the call to ::glow:optimize(...)

You can then convert these .dot files to png's and see the optimization.