Remove vtable from node
Remove the vtable from the Node class. The tricky part here is that in order to
remove the vtable from the destructor we need to be able to destroy derived
classes that have some state. The mechanism that we use here is the static
dispatch in ~Node(). To ensure the correct destruction of nodes we need to call
Graph.eraseNode().