GraphOptimizer: Sink transpose operations below sigmoid/tanh operations
Sigmoid and Tanh are types of activation functions commonly used in
Machine Learning.  In this commit, we want to optimize our graph that
we generate to sink the Transpose nodes below the Sigmoid and Tanh nodes.

This is similar to an already existing sink optimization that we have with
Transpose and RELU nodes. This will allow us to potentially merge these
transpose operations with other transpose operations.

Also included are some tests for these two optimizations.

Ori Shalev originally wrote many of the tests for our optimizations and
I want to point out a very useful debugging technique he mentioned
for debugging our optimizations is to add G.dumpDAG("before.dot") and
G.dumpDAG("after.dot") around the call to ::glow:optimize(...)

You can then convert these .dot files to png's and see the optimization.