Introduce caching and non-caching versions of PoolMax and SoftMax
Until now, the compiler had only caching versions of PoolMax and SoftMax. In case of inference, these caches were not used at all and just increased memory consumption and execution time.

With this change, the compiler replaces caching instructions by their non-caching versions if it can prove that caches are not used.