New and more correct version of buffer sharing
This is a new version of buffer sharing which subsumes old versions of both buffer sharing and copy propagation. It unifies both optimizations and it fixes some bugs discovered in the old, simpler implementation.

The algorithm is using live intervals. The overall idea is that it is fine to combine storage for two live intervals if they do not overlap.