[jit] Support specialization of functions for constant arguments
This optimization creates specialized functions from general-purpose function provided a set of constant arguments. For example, it would generate specialized versions of convolutions for specific inputs and filter dimensions. Having such specialized functions allows for better optimizations by the LLVM backend, as it has more information about the values of function arguments.

For now, this optimization is only performed for the batched matrix multiplication and convolutions, but it is easy to support any other operations.

There are two command-line flags controlling this optimization:
`-jit-specialize-dims` - creates specialized functions for operations with constant dimensions. This option is enabled by default.
`-jit-specialize` - creates specialized functions for operations with constant arguments. This option is disabled by default.