[jit] Support specialization of functions for constant arguments
This optimization creates specialized functions from general-purpose function provided a set of constant arguments. For example, it would generate specialized versions of convolutions for specific inputs and filter dimensions. Having such specialized functions allows for better optimizations by the LLVM backend, as it has more information about the values of function arguments. It also significantly improves the JIT compilation speed, because the generated LLVM module consists of multiple functions of smaller sizes instead of having only a huge `main` function. The performance of the generated code does not show any degradations.

For now, this optimization is only performed for all functions invoked from main. In the future, we may introrduce more complex logic for making this decision. It could be based in the number of invocations of a function, number of its arguments, its code size, etc.

There is a command-line flag controlling this optimization:
`-jit-specialize` - creates specialized functions for operations with constant arguments. This option is enabled by default.