[Graph] Change some of the APIs from reference to pointer to make them more consistent.
Some of our APIs need to be pointers (for nullability, compatibility with C wrappers, convenience, etc.) and we were using references and pointers randomly. This commit goes in the direction of adopting the LLVM rule of using pointers for referring to functions and let the Module manage their lifetime. This is also consistent with the way we use Nodes in the graph.