Add support for serializing bundles
Each backend can now implement a `save` method to generate a bundle. As of now, only the JIT backend supports it.

To generate a bundle, use e.g. the following command-line:
loader tests/images/cat_285.png -image_mode=0to1 -d=resnet50 -jit -emit-bundle directory_path

This command would generate two artifacts: the code is serialized into the resnet50.o file and the weights are serialized into resnet50.weights file in the directory specified by the `-bundle-output-dir` parameter.

The generated code is a usual object file which you can link with any client code.

The code exposes an entry point, which has the same name as the bundle, i.e. resnet50 (the name is taken from the model being processes). This entry point has the following function signature and takes the pointers to the initialized memory allocated for constant and mutable weight variables:
`extern "C" void resnet50(uint8_t *constantWeightVarsBaseAddr, uint8_t *mutableWeightVarsBaseAddr, uint8_t *activationsBaseAddr);`

Another exposed symbol has a name which looks like the name of the bundle, but with a suffix `_config` at the end. In the example above, it would have a name `resnet50_config`. This is a structure of the following form, which contains the information about the sizes of the memory blocks for activations, constant weights variables and mutable weights variables:
```c++
struct BundleConfig {
size_t constantWeightVarsMemSize;
size_t mutableWeightVarsMemSize;
size_t activationsMemSize;
};
```

It is a responsibility of the client to properly prepare and initialize the memory blocks for the activations, constant weights variables and mutable weights variable before calling the exposed entry point.