Add an example that creates a standalone bundle for lenet_mnist
lenet_mnist_standalone.cpp is a small wrapper that links with and invokes the bundle auto-generated by means of the loader with the -emit-bundle parameter. All of the logic for images loading is essentially taken from the loader. The bungle itself is serialized into files lenet_mnist.o and lenet_mnist.weights

The standalone bundle can be invoked by using the following command-line:
lenet_mnist_standalone image_file

The commit also provides a documentation file describing how to generate and use bundles in general and providing also a step-by-step explanation of the lenet_mnist example.