[Quantization] Fix the end-to-end quantization unittest.
The unit test had some problems that made it fragile (weights were
shared, only 2 floats were compared, etc), and broken, because when we
increased the number of elements the test failed.

I rewrote the test to include two identical independent networks. One
network is used for profile collection and other other for quantization.