[JIT] Add support for the debug_print instruction
This feature is supposed to be used only for debugging. It is very useful when you need to find differences between the execution of the network model by the interpreter and by the JIT, as you can compare two execution longs step-by-step.  I used it to debug JIT compilation problems with PTB.