[JIT] Make specializer more conservative
Fixes a bug uncovered by recent libjit_copy related commits. The new pattern used a bitcast instruction, which was not caught as a non-constant. As a result, a non-constant argument was used to produce a specialization. This in turn resulted in creating specialized functions referring to non-constant llvm::Values defined in a different function.