[Quantization] Fix the shift scale calculation in quantize32To8
[Quantization] Two bugs in the scale calculation. First, if post is zero
we have undefined behavior in the code that calculates the shift.
Second, if for some reason the input scale is large, for example, the
case of convolution bias, we need to limit the shift amount.

With this change the quantized Resnet50 works in the CPU backend
(without floating point ops inside the convolution.)