[JIT] Support the emission of debug information
There is a new command-line option `-g`, which can be used to emit debug information.
If this option is used, it is possible to debug the generated AOT bundles using a debugger like LLDB.

When you stop the execution, the debugger would show you the name of the current function and its arguments, e.g. buffer pointers, number of dimensions, etc. At the moment, the names of the arguments are not very user-friendly and do not reflect their semantics. They are called arg1, arg2, etc. I'll try to improve it in the future. But even with this, you can use the usual debugger commands to inspect the values, e.g. `print arg1[10]`, to inspect the 10th element of the buffer.

Stepping does not work properly, because there is no textual representation of a network model at the moment. But I'm thinking about adding an even more advanced debugging support in the future, so that one can execute Glow low-level IR step-by-step using a debugger. To do that, we just need to serialize the Glow low-level IR into a text file and use the line numbers referring to it.