Implement The  DKKC8 cpu convolution.
This commit adds a new kind of cpu-specific convolution that operates on
filter weight data in a non-standard format.  The default format is
DKKC, where D is the output depth of the filter and C is the input
channel, and K is the kernel size. This commit changes the data layout
to [D/8, K, K, C, 8].  We pre-swizzle the data in the weights to make
the access pattern more efficient.

This commit creates a new target-specific node and instruction and
detects profitable patterns. This provides a significant performance
boost on Resnet50.