Backends: support older CMake
For older cmake, we cannot use the generator expression.  Instead, abuse the
command line and properties to accomplish our goal.  Use a pre-link command
(which is chained to the linker invocation) to actually link the object files
using a custom rule and use the comment leader (`#`) to avoid the actual shared
link.  This allows us to generate the AOT library with a custom linker in a
manner that allows us to use an older cmake.