CPU: adjust build rules to allow building with GCC
The CPU Runtime requires clang as we need to emit LLVM bitcode to perform the
LTO'ed AOT for the JIT.  Chain a pre-compile rule to override the compiler
invocation.  This allows the rest of glow to be built with GCC as long as clang
is provided on the system as well.