[opencl] Add support for profiling OpenCL kernels
Collecting statistics about the executions of specific OpenCL kernels is rather difficult, as usual profiles are not quite capable of doing it. But this information is very important for understanding what needs to be improved and optimized. Therefore, this patch add some basic supporting for collecting this kind of statistics:

- It introduces a new command-line option -opencl-profile which enables the profiling
- When this option is used, the following information is collected:
   - Duration of each launched OpenCL kernel.
   - Total duration of each kernel during the execution of the graph. This is the sum of durations for launches of the same kernel.