[AOT] Add an example that creates a standalone AOT bundle for resnet50
resnet50_standalone.cpp is a small wrapper that links with and invokes the bundle auto-generated by means of the loader with the -emit-bundle parameter. All of the logic for images loading is essentially taken from the loader. The bungle itself is serialized into files resnet50.o and resnet50.weights

The standalone bundle can be invoked by using the following command-line:
resnet50_standalone image_file

The commit also provides a step-by-step explanation of the resnet50 example.