[CPU] Implement a gemm-like unrolled convolution for 1x1 filters.
This commit implements a gemm-like unrolled convolution for 1x1 filters.
We optimize things by loading data from the filter buffer once and use
it for multiplication of two input values.

On Resnet50 this gives only a small speedup (of about 5)%, because most
of the time is spent in convolutions with non-1 filter size, which are
also computationally more demanding.