Introduce a helper function to remove an IR instruction by providing the iterator referring to it
This allows for a quick removal without performing a linear search for the instruction in the instruction list which can be rather long.

Use this new function in the IROptimizer to speed-up the compilation. When rewriting the code to use this new API, make use of the fact that an instruction created by means of IRBuilder::createXYZ functions is always added at the end of the instructions list and thus can be easily accessed using `instrs.back()`.

Also avoid performing two scans of the instructions list in hoistDealloc. Collect all deallocation instructions during the first scan.