[ir-optimizer] Improve shareBuffers to better handle instructions with tensorview operands
shareBuffers would sometimes miss the opportunity to share buffers between operands of an instructions if some of these operands were tensorviews. This should be fixed now.

After more aggressive sharing of buffers involving tensorviews it may happen that some of the tensorviews are  used before they are defined. This is fixed by a sinkTensorViews helper, which sinks tensorviews right before their first use, similar to sinkAllocas.