[opencl] Introduce macros to define data-parallel operations and define operations in their terms
This is similar to what we do in libjit.cpp for the CPU backend.
Since OpenCL supports vector types, the macros generate multiple versions of a data-parallel operation for vectors of sizes 1, 8 and 32 and the OpenCL backend will pick the appropriate kernel depending on the concrete dimensions of the operations to be performed.