[CPU] Changed the heart of heart of the matmul operation to perform 12 multiplications instead of 10.
Before this change we had data-dependent chains of length 4. This means that our FMAs were only 80% utilized. Changing the block from 4*2 to 3*4 we changed the memory-compute ratio. We now perform 12 FMAs for 7 loads (compared to 8 FMAs for 6 loads).

This gives a significant speedup on some matrix multiplications. For example, we now go from 80 Gflops to 10 Gflops on [768 x 768] * [768 x 768].