[opencl] Add implementation of the oclconvolution instruction
This implementation uses a parameterized convolution kernel heavily based on the kernels produced by libDNN.
The kernel is parameterized by means of macro-definitions, which define such properties like size of the filter, padding, stride, sizes of workgroups, etc.
The OpenCL backend produces specialized convolution kernels based on it.

The kernels use some typical GPU tricks to improve performance, e.g.:
- Using __local memory to load input and kernel tiles and share them between all threads in a work group to reduce the amount of accesses to the global memory
- Using  local variables of vector types to speed up computation in each thread and to accumulate results
- Explicit loop unrolling and pragmas controlling the loop unrolling

These kernels are much faster than the default NHWC convolution kernels that were used so far. The performance with a batch size of 1 approaches the performance of the CPU convolutions. And with batches of size bigger than 2 it surpasses the the performance of the CPU backend e.g. on resnet50.