[Graph] Allow nodes in different functions to have the same names.
Allow nodes in different functions to have the same names. This change allows us to clone functions and keep the original names. This is useful for a number of use cases. First, we can perform quantization on the cloned functions. Second, preserving node names is useful for multi-versioning compilation for different resources on the device and for different batch sizes. Finally, this allows us to remove some of the workarounds that we have around function cloning.