Add unittest for GRU quantization
createGRUForQuantization combines the GRU specification from /examples/fr2en.cpp
with the random initialization from the quantization end2end test to build a
graph with 2 recurrent units. The weight vector remains the same in each GRU while
the set of "words" and "embeddings" change. A TopKNode which cuts the hidden
unit in half is added to test quantization of this node.
The graph is the quantized with the same steps as the end2end test.

Addresses #985