Add a new example program: char-rnn.
The char-rnn program is based on a blog post by Andrej Karpathy: "The
Unreasonable Effectiveness of Recurrent Neural Networks". This is a
small program that reads text from a file and generates random text
based on the input. In each step the network is asked, given the current
buffer of N chars, what is the next char? After a few minutes of
training interesting texts come up.

This is a text program that's designed to stress the Glow compiler
optimizer. There are a bunch of unrealized opportunities for
optimizations, for example horizontal fusion of softmax and matrix
multiplications, buffer optimizations, etc.