[CPU] Do not emit any warnings or diagnostics when JITting
By default, LLVM would emit some diagnostics, remarks, etc. It is fine for a static compiler, but not necessary for a JIT. Let's disable it by providing a dummy diagnostics handler, that does not emit anything. In particular, this allows us to get rid of the annoying "cannot vectorize" warnings.

The LLVM APIs for setting the diagnostics handlers have changed in the LLVM 6.0, therefore we have to use a slightly different code for LLVM versions before and after LLVM 6.0.