[cmake][test] Add a add_glow_test function and 'check' target
Introduce the add_glow_test function that on top of adding the test
to the list of tests to be run with ctest, it will register the
dependencies on the actual executable.
Thanks to this information, the new 'check' target can re-build
the test executables if needed and then run the test.

In other words, the combination of those two additions act like
`ninja all && ninja test`, except we don't have to think about it.