[ir-optimizer] Compilation speed-up: Use instruction iterators instead of instruction pointers when calling moveInstruction and eraseInstruction
Recent testing with big NN models has uncovered some compilation speed issues caused by the IR optimizer.

One of the reasons is that the versions of moveInstruction and eraseInstruction functions that take a glow::Instruction pointer as a parameter were called. These versions are very inefficient as they need to first find the instruction in the linked list representing the instructions list.

This commit switches to using instruction iterators instead of instruction pointers. This solves all the compilation speed issues related to moveInstruction and eraseInstruction.