[libjit] Copy the offsets vector into a local buffer to signal that the buffers don't alias.
When we decide not to specialize the offsets buffer then the optimizer is having a hard time figuring out that the offsets (that are passed as a pointer to some external memory) don't alias with the outputs of this method. By copying them locally we signal to the optimizer that the memory does not alias. After this change the code is nicely optimized, even after unspecializing the offsets buffer.