[debug-info] Strip any debug info from the loaded Glow's library (libjit) if the emission of debug info is not enabled
No debug information is going to be emitted for the code generated from the Glow IR if  the emission of debug info is not enabled.

But any debug information from the Glow's library (e.g. libjit) should be stripped as well. Let's strip the debug info as early as possible, so that the LLVM's optimizer does not need to spend any time to preserve the debug info during optimizations.

It also makes the LLVM IR dump much smaller.