[debug-info] Legalize names of all instructions
Names of instructions inside IRFunctions are always uniqued. But they were not legalized and thus they could not be used in the debugger, when debugging Glow networks, because debuggers expect legal C identifiers as variable names.

This change follows the same approach as Graphs and Modules and legalizes the names of instructions.