[debug-info] Do not explicitly emit debug info for specialized functions
There is no need to explicitly emit a debug info for the specialized function. If the original function had it, the cloner would have automatically copied it into the specialized function. And if the original function did not have any debug info, then its specialization should not have any debug info either.