[debug-info] Do not emit any debug info for libjit
The debugging of Glow networks should happen at the Glow IR level. Being able to step into heavily optimized libjit kernels is not very helpful, because debuggers usually have a rather weird behavior inside optimized functions. Moreover, the source code of libjit is likely to be unavailable in production use and thus the debugging would not be possible anyways for libjit's functions.