Make sure that use-lists of Nodes only contain uses by other Nodes
Until now, NodeValue was used to refer to the results of Nodes. But unfortunately, every NodeValue would add itself to the use-list of a Node it references, which is rather expensive. In many cases, NodeValue objects are created as temporary objects and it makes no sense to consider them to be uses.

We want to count only references from Nodes to the results of other Nodes as uses. To achieve that this commit introduces a new type NodeValueHandle. It is used instead of NodeValue inside Nodes to refer to the results of other nodes. All NodeValueHandles add themselves as uses to the use lists. And the former NodeValue is changed in such a way that it does not register itself in the use-lists anymore.

To ensure that NodeValueHandle cannot be used in the code outside of Node and its derived classes some kind of a type system  based access control is introduced by means of introducing a "namespace" PrivateNodeTypes, whose members can be accessed only by Node and its derived classes.