[GraphOptimizer] Fix batchNorm optimization
The batchNorm optimization used to assume that the inputs of the related
convolution node were all variables. This is not enforced anywhere and may
actually be wrong.

The fix consists in checking that the inputs of the convolution node are indeed
variables. Moreover, we check that these variables are used only by this
convolution node before doing the transformation. Otherwise we will incorrectly
change the inputs of the other users.

This fixes issue #1178 for the graph optimizer.